Refactoring Registry.go (#44)
* Moving things around


* Fixing some tests


* Rebasing


* Another test


* More tests


* Another test
