+++ admiral/pkg/controller/admiral/configmap.go
@@ -1,10 +1,6 @@
 package admiral
 
 import (
-	"errors"
-	"github.com/istio-ecosystem/admiral/admiral/pkg/controller/common"
-	"github.com/sirupsen/logrus"
-	"gopkg.in/yaml.v2"
 	"k8s.io/api/core/v1"
 	metaV1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/client-go/kubernetes"
@@ -84,42 +80,7 @@ func (c *ConfigMapController) GetConfigMap() (*v1.ConfigMap, error) {
 }
 
 func (c *ConfigMapController) PutConfigMap(newMap *v1.ConfigMap) error {
-	err := validateConfigmapBeforePutting(newMap)
-	if err != nil {
-		return err
-	}
-	_, err = c.K8sClient.CoreV1().ConfigMaps(c.ConfigmapNamespace).Update(newMap)
+	_, err := c.K8sClient.CoreV1().ConfigMaps(c.ConfigmapNamespace).Update(newMap)
 	return err
 }
 
-func GetServiceEntryStateFromConfigmap(configmap *v1.ConfigMap) *common.ServiceEntryAddressStore {
-
-	bytes := []byte(configmap.Data["serviceEntryAddressStore"])
-	addressStore := common.ServiceEntryAddressStore{}
-	err := yaml.Unmarshal(bytes, &addressStore)
-
-	if err != nil {
-		logrus.Errorf("Could not unmarshal configmap data. Double check the configmap format. %v", err)
-		return nil
-	}
-	if addressStore.Addresses == nil {
-		addressStore.Addresses = []string{}
-	}
-	if addressStore.EntryAddresses == nil {
-		addressStore.EntryAddresses = map[string]string{}
-	}
-
-
-	return &addressStore
-}
-
-func validateConfigmapBeforePutting(cm *v1.ConfigMap) error {
-	if cm.ResourceVersion == "" {
-		return errors.New("resourceversion required") //without it, we can't be sure someone else didn't put something between our read and write
-	}
-	store := GetServiceEntryStateFromConfigmap(cm)
-	if len(store.EntryAddresses) != len(store.Addresses) {
-		return errors.New("address cache length mismatch") //should be impossible. We're in a state where the list of addresses doesn't match the map of se:address. Something's been missed and must be fixed
-	}
-	return nil
-}
