+++ admiral/pkg/test/types.go
@@ -1,10 +1,7 @@
 package test
 
 import (
-	"github.com/istio-ecosystem/admiral/admiral/pkg/controller/common"
-	"gopkg.in/yaml.v2"
 	k8sCoreV1 "k8s.io/api/core/v1"
-
 )
 
 type FakeConfigMapController struct {
@@ -19,15 +16,4 @@ func (c *FakeConfigMapController) GetConfigMap() (*k8sCoreV1.ConfigMap, error) {
 
 func (c *FakeConfigMapController) PutConfigMap(newMap *k8sCoreV1.ConfigMap) error {
 	return c.PutError
-}
-
-func BuildFakeConfigMapFromAddressStore(addressStore *common.ServiceEntryAddressStore) *k8sCoreV1.ConfigMap{
-	bytes,_ := yaml.Marshal(addressStore)
-
-	cm := k8sCoreV1.ConfigMap{
-		Data: map[string]string{"serviceEntryAddressStore": string(bytes)},
-	}
-	cm.Name="se-address-configmap"
-	cm.Namespace="admiral-remote-ctx"
-	return &cm
 }
